requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
